model/requirements.txt
flask
pandas
numpy
psycopg2-binary
scikit-learn
xgboost
python-dotenv
shap
seaborn
